chore: rename to follow convention
