+++ app/components/messaging/identity-verification-notice.js
